requirements.txt
fastapi
uvicorn
motor
pydantic[email]
pydantic-settings
python-dotenv
python-multipart
passlib[bcrypt]
python-jose[cryptography]
bcrypt==4.0.1
email-validator
pandas
openpyxl
httpx